Distance to walk

+++ src/Menu.h
@@ -16,9 +16,11 @@ using namespace std;
 struct choosingWay
 {
     int howToChooseRoute;
+
     bool goOnFoot = false;
     bool nightTime = false;
-    int metresToWalk;
+    float metresToWalk;
+
     unordered_set<string> hiddenLines;
     float startLat =0.0, endLat=0.0, startLong=0.0, endLong=0.0;
     string startStop=" ", endStop =" ";
